ingest-drive-files/requirements.txt
attrs==23.1.0
bcrypt==4.1.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Cython==0.29.36
distro==1.8.0
docopt==0.6.2
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
jsonschema==3.2.0
oauthlib==3.2.2
packaging==23.2
paramiko==3.3.1
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.20.0
python-dotenv==0.21.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
texttable==1.7.0
uritemplate==4.1.1
urllib3==2.1.0
websocket-client==0.59.0
